RockLib.Messaging.Http version 1.0.0-alpha05
Fixes a bug in path regex of http receiver, where it would handle a request with extra path elements instead of returning a 404. For example, if the receiver's path is "/api/{api_version}" and an incoming request had a path of "/api/v1/extra_stuff", the response should be a 404.