RockLib.Messaging version 2.3.0
Always register a lookup delegate from the `AddSender`, `AddTransactionalSender`, and `AddReceiver` extension methods. The previous version would only register the lookup delegate if all of the messaging services of that type were registered as singleton. Also, for each implementation of lookup delegate, if a transient/scoped service is created but not needed, dispose it.