SQS - Reset visibility timeout on rollback
Add the ability to reset a sqs message's visibility timeout when the
message is rolled back.

When rolling back a message, the intent is generally to retry the
message. Currently, the message will not be retried until the
visibility timeout expires. While this may be suitable in some cases,
it can cause added queue pressure in fifo queues since any messages
still in the queue with the same group id would be held up. Terminating
the visibility timeout on rollback allows the message to become
immediately available for queue consumers.